Refactor math stuff into new module
